atlantisdiagnostics: Tools for calibrating Atlantis ecosystem models

Tools to aid in diagnosing issues in a parameterized Atlantis Ecosystem model

Details

Determine functional group persistence over time

Determine functional group stability over time

Determine functional group biomass (relative to historic data) lie within reasonable bounds

Determine distribution of biomass over age classes

Determine maximum size (length and weight) of species relative to observations

References

To learn more about using atlantisdiagnostics, start with the vignette: browseVignettes(package="atlantisdiagnostics") or click the index below

See also

Author

Maintainer: Andy Beet andrew.beet@noaa.gov (ORCID)

Authors:

Sarah Gaichas Sarah.Gaichas@noaa.gov

Joeseph Caracappa joseph.caracappa@noaa.gov

Robert Gamble Robert.Gamble@noaa.gov

Ryan Morse ryan.morse@noaa.gov (ORCID)