Test for functional group reasonability

Inspects each time point. If at any point in time a species falls below or rises above predefined bounds it is flagged. The term "reasonable" is based on survey data when available.

diag_reasonability(
  modelBiomass,
  initialYr = 1964,
  speciesCodes = NULL,
  realBiomass,
  nYrs = NULL,
  surveyBounds = c(1, 1),
  initBioBounds = c(0.5, 10),
  plot = F
)

Arguments

modelBiomass
A data frame. Total biomass of all groups over time, read in from Atlantis ...BioInd.txt output using atlantisom::load_bioind.

initialYr
Numeric Scalar. Year in which the model run was initiated. (Default = 1964)

speciesCodes
Character vector. A vector of Atlantis species codes in which to test for reasonableness (Default = NULL, uses all species found in modelBiomass. Species codes should be a subset of the Atlantis species codes

realBiomass
A data frame. biomass time series (from assessments, stock SMART or otherwise) for species. CURRENTLY IMPLEMENTED USING SURVDAT DATA ONLY realBiomass should be in long format with column labels (YEAR,variable,value,Code,Species,isFishedSpecies) Biomass units should be in metric tonnes

nYrs
Numeric scalar. Number of years from the end of the time series for which reasonableness is checked. (Default = NULL, entire time series is used)

surveyBounds
Numeric vector. Size of 1x2 containing the values in which to multiple lower and upper bounds of observed data. For example (Default = c(1,1)) indicating use of min and max of observed biomass

initBioBounds
Numeric vector. Size of 1x2 containing lower and upper bound proportions used to scale initial biomass. This is used for groups/species that dont have surveys

plot
A logical value specifying if the function should generate plots or not. (Default = F). NOT YET IMPLEMENTED

Value

Returns a data frame of species

species

The common name of the species/functional group as described in Atlantis input file

code

Atlantis Code for species/functional group

initialBiomass

Starting value of Biomass for species/functional group. From model output

minBiomass

The smallest value of biomass observed in the run

maxBiomass

The largest value of biomass observed in the run

propInitiBio

The maxBiomass as a proportion of the inintial biomass

t1

The first year reasonableness was not met

tn

The last year reasonableness was not met

nts

The total number of years that reasonableness was not met

modelSkill

Measure of the goodness of fit

nObs

number of observations used in the goodness of fit

pass

Boolean. Indicates if the Atlantis group passed reasonability

test

Character. Indicate how pass was determined. If the Atlantis species is fished and the species is a stand alone species in the model then survey data is used to assess reasonability. Otherwise deviations from initial biomass is used

Examples

if (FALSE) {
}