Test for model stability

diag_stable determines whether the model reaches steady state for the last ~20 years of an unperturbed, unfished ~100 year run.

diag_stable(atBtxt, fgs.names, runpar, plot = TRUE)

Arguments

atBtxt
A dataframe of total biomass of all groups over time, read in from Atlantis ...BioInd.txt output using atlantisom::load_bioind.

fgs.names
A vector of species names

runpar
A list of Atlantis run parameters read in from the .xml file using atlantisom::load_runprm

plot
A logical value specifying if the function should generate plots or not. The default is TRUE.

Value

Returns a dataframe of crashed species with columns "Failed" of species names and "nYrs0B" with number of years having 0 biomass. Also plots all species trajectories with orange (positive) or purple (negative) trend lines for unstable groups over the last 20 years unless plot=FALSE.

Details

This test evaluates the last 20 years of the input run for a significant slope using the geom_gls function from the package ecodata at https://github.com/NOAA-EDAB/ecodata. The function diag_stable also checks for persistence and flags groups that have gone to 0 biomass--they don't count as stable. It takes the same inputs as diag_persist with the addition of the runpar object, and defaults to plotting, which can be turned off with plot=FALSE in the function call.

Author

Sarah Gaichas

Examples

if (FALSE) {
# Usage
diag_stable(testdiag$atBtxt, testdiag$fgs.names, testdiag$runpar)
}