Test for model stability

diag_stability determines whether the model reaches a steady state over the last n years of a run. Stability is loosely defined as a species/groups biomass reaching a stable level (measured as having no trend in a defined time range).

diag_stability(modelBiomass, speciesNames, nYrs = 20, sigTest = 0.05, plot = F)

Arguments

modelBiomass
A data frame. Total biomass of all groups over time, read in from Atlantis ...BioInd.txt output using atlantisom::load_bioind.

speciesNames
Character vector. A vector of species names in which to test for stability. (Default = NULL, uses all species found in biomass)

nYrs
Numeric scalar. Number of years from the end of the time series that stability must occur. (Default = 20 years)

sigTest
Numeric Scalar. alpha level to use to test for slope significance (Default = 0.05)

plot
Logical. Specifying whether the function should generate plots or not. (Default is F).

Value

Returns a data frame of all species and how they measure up against the stability criterion

species

The common name of the species/functional group

slope

Double. The value of the slope parameter (time)

pValue

Double. The p-value associated with the slope parameter (time)

pass

Logical. Does the species/group pass the test for stability

Details

Formally the following model is fit to the last n years of the run:

$$biomass_t = \mu + \beta.t + \epsilon_t where \epsilon_t ~ IID N(0,\sigma^2)$$

where null hypothesis, $$H0:\beta=0$$

Examples

if (FALSE) {
# Declare paths to files required
biol.file <- "neus_outputBiomIndx.txt"
file_fgs <- "neus_groups.csv"
# use atlantisom to read them in
fgs <- atlantisom::load_fgs(inDir,file_fgs)
biomass <- atlantisom::load_bioind(outDir,biol.file,fgs)

# Perform stability test on all species/groups using the last 20 years of the run
diag_stability(biomass)

# Only perform test on herring and white hake.
# Require stability over the last 10 years of the run and and use alpha = 0.1 as tests significance level
diag_stability(biomass, speciesNames=c("Herring","White_Hake"), nYrs = 10, sigTest = 0.1)
}